+++ packages/create-svelte/template-additions/.eslintrc.ts.cjs
@@ -1,0 +1,15 @@
+module.exports = {
+	root: true,
+	parser: '@typescript-eslint/parser',
+	extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+	plugins: ['svelte3', '@typescript-eslint'],
+	ignorePatterns: ['.eslintrc.js'],
+	overrides: [{ files: ['*.svelte'], processor: 'svelte3/svelte3' }],
+	settings: {
+		'svelte3/typescript': require('typescript')
+	},
+	parserOptions: {
+		sourceType: 'module',
+		ecmaVersion: 2018
+	}
+};