+++ packages/kit/src/runtime/client/router.js
@@ -1,10 +1,5 @@
 import { find_anchor, get_base_uri } from './utils';
 
-/** @param {MouseEvent} event */
-function which(event) {
-	return event.which === null ? event.button : event.which;
-}
-
 function scroll_state() {
 	return {
 		x: pageXOffset,
@@ -73,7 +68,7 @@ export class Router {
 		addEventListener('click', (event) => {
 			// Adapted from https://github.com/visionmedia/page.js
 			// MIT license https://github.com/visionmedia/page.js#license
-			if (which(event) !== 1) return;
+			if (event.button || event.which !== 1) return;
 			if (event.metaKey || event.ctrlKey || event.shiftKey || event.altKey) return;
 			if (event.defaultPrevented) return;
 
@@ -144,14 +139,13 @@ export class Router {
 
 		const path = url.pathname.slice(this.base.length) || '/';
 
-		const routes = this.routes.filter((route) => route.pattern.test(path));
+		const routes = this.routes.filter(([pattern]) => pattern.test(path));
 
 		if (routes.length > 0) {
-			return {
-				routes,
-				path,
-				query: new URLSearchParams(url.search)
-			};
+			const query = new URLSearchParams(url.search);
+			const id = `${path}?${query}`;
+
+			return { id, routes, path, query };
 		}
 	}