+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-svelte
 
+## 2.0.0-next.42
+
+### Patch Changes
+
+- a52cf82: add eslint and prettier setup options
+
+## 2.0.0-next.41
+
+### Patch Changes
+
+- 8024178: remove @sveltejs/app-utils
+
 ## 2.0.0-next.40
 
 ### Patch Changes